[memory] Fix level bug in PressureNotifier
Make sure that the memory pressure level a Watcher of the memorypressure
service sees is one of Normal, Warning, Critical. This means that if the
current system pressure level is Imminent-OOM, it needs to be converted
to Critical for the Watcher.

This was a bug only on the callback and register paths. On the normal
level change path, we take care to not notify with the Imminent-OOM
level.

Test: pressure_notifier_unittest
Fixed: 76980
Change-Id: I9621cb0770e5249b3e228d46ac2b6123a50c13d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/529942
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: David Moore <davemoore@google.com>