[codec] Create component context before use
The use_media_decoder shell utility was crashing in main because it was
attempting to use the sys::ComponentContext prior to creation. This CL
splits the creation of the ComponentContext from serving the output
directory so that we may access incoming services without publishing our
outgoing until ready.

Change-Id: I29b69d9b68b681849b7bf33a67dd3b29581fd6c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530909
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: RJ Ascani <rjascani@google.com>