[bt-init] Define temporary v2 bt-init component.
Define a manifest and separate binary for a v2 bt-init component. To
support the optional presence of bt-rfcomm, use CF APIs to attempt to
open the Profile service of the bt-rfcomm child component, falling back
to bt-gap's Profile service if opening bt-rfcomm fails.

The manifest's child URLs are absolute paths to production components,
which will not allow them to be packaged with bt-init for hermetic
integration testing. As such, this manifest will never be used as-is.
We will convert it to a CML shard which parameterizes its children (see
fxbug.dev/74547). It exists in this CL solely for making the review
process simpler.

Bug: 72282
Test: `fx test bt-init-unittests-v2` to verify logic handling optional
      RFCOMM presence and maintain unit test coverage from v1. TODO -
      this is a step towards unblocking integration test coverage in
      CQ (fxbug.dev/53915), so this component will be thoroughly
      integration-tested before it is ever used in production. Also, we
      will write a manifest smoke test for this manifest once converted
      to a CML shard.
Change-Id: Id6f08ef192c0e0cf75668fef29ad9c4ddb415449
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504726
Commit-Queue: Lucas Jenkins <lucasjenkins@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>