[ffx] add some CI target awareness to self-test
In CI, the leased target nodename is passed in via FUCHSIA_NODENAME, and
this is the target that we want to interact with. In other environments
where FUCHSIA_NODENAME is not set, attempt to discover a target in a
product state and return that nodename.

Fixed: 76281
Change-Id: I721089630f205c7312c5c260c7ad5c6ec270ca92
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530911
Commit-Queue: James Tucker <raggi@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>