[testrunner] Parse out kernel logs from test output.
Kernel logs are already getting printed to stdout from botanist. Bringup
tests are run on serial so we print the serial output which contains the
kernel logs and test output to stdout. This causes the kernel logs to
get printed twice which causes some tefmocheck cases to fail which look
for strings in the swarming output.

Bug: 74703
Change-Id: Iff02ba403b0190e5f72e95865d1c7580934be218
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528421
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>