[async-utils] Fix EventWaitResult drop panic.
Do not panic if the list of wakers does not contain the current waker.
The waker can be missing if the Event was dropped or signaled, waking
the waiters and removing them, and the Result was not repolled before
being dropped.

Test: Added unit test for this behavior
Bug: 76728
Change-Id: Ibc30fc0f5c4d06914d350798b677582bc03255a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530511
Fuchsia-Auto-Submit: Jeff Belgum <belgum@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>