[sys] Lift pkg-resolver into v2 under core
This change swaps the load-bearing instance of pkg-resolver to live
as a v2 component under the core realm instead of being a v1 component
run under appmgr.

pkg-resolver exposes five protocols:

* fuchsia.pkg.FontResolver, which is now only consumed by v2 components
* fuchsia.pkg.PackageResolver, fuchsia.pkg.PackageResolverAdmin,
  fuchsia.pkg.RepositoryManager, and fuchsia.pkg.rewrite.Engine, which
  all have remaining v1 clients and thus are wired back into appmgr

There's another wrinkle in here, which is that sysmgr previously had
explicit checking to ensure that the fuchsia.pkg.PackageResolver service
was listed in some `sysmgr_config`'s `update_dependencies` section, but
now that sysmgr is no longer responsible for spawning pkg-resolver, that
check should be removed.

However, a number of tests still rely on the (implied) behavior of
sysmgr not making use of the system-wide PackageResolver.  To avoid
breaking those tests, we specify a nonsense service in
`update_dependencies` for those sysmgr instances, which will cause those
sysmgr instances to never attempt to use PackageResolver (and instead
fall back to using the base resolver).

Testing: `fx set core.qemu-x64 ; fx build ; fx emu -N` and then running
over ssh a program (ls) that is not available in the base package set
and seeing that it resolves and runs correctly.

Change-Id: I58ba8ee39046dc887335390f4f7e6d8bc8737143
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/527082
Reviewed-by: Dan Johnson <computerdruid@google.com>
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>