[ffx][test] Create ffx test
Create 'ffx test' as a replacement for 'ffx component test'. For now,
this duplicates the logic in 'ffx component test' and adds it to
'ffx test run', with the exception of listing test cases, which is put
in 'ffx test list'. 'ffx test result' is added for new commands that
will allow a developer to inspect the results of previous runs and will
be filled in during subsequent commits.
'ffx test' is marked experimental for now.

Bug: 75133

Change-Id: Ib7b27ddc1715859be5d3c4ad809173666a0fe7e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528289
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Matthew Boetger <boetger@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Satsuki Ueno <satsukiu@google.com>