[SetUI] Check channel state before messaging on exit.
A number of places communicate back to a parent component
during teardown. For example, messengers message the
MessageHub during drop to indicate deletion. It is possible that
the parent component is no longer listening due to an immediate
exit of their event loop as part of a larger teardown. The send
will fail in these cases.

This changelist prevents sending to a closed channel by checking
first.

Fixed: 77024
Test: fx test -o setui_service_tests
Change-Id: Iaa819a24723fcdd7763b7c78fc7ddf8e982b165b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/531041
Reviewed-by: William Xiao <wxyz@google.com>
Commit-Queue: Bryce Lee <brycelee@google.com>