[SetUI][FIDL] Revise interface handling.
This changelist introduces a new approach for  registering fidl
interfaces with the setting service. Previously, the registered
interfaces were inferred by the setting types specified in the
service configuration. Each interface was explicitly handled in
the general service bringup. This relationship is flipped - the
presence of an interface should define the inner components
the service should bring up. To correct this behavior, interfaces
are now explicitly defined as part of the configuration (note: we
map the setting type definitions to interfaces until the
configuration file schema can be updated).

As part of this change, a new way for registering components with
inbound communication is introduced. Registrant provides a
framework for bringing up such interfaces beyond FIDL. Registrants
allow capturing details such as dependencies. Dependencies are
expressed in terms of Entities, which identify components such as
setting handlers. The setting service brings up inner copmonents
based on the union of dependencies for its registrants.

Fixed: 75986
Test: fx test -o setui_service_tests
MULTIPLY: setui_service_tests
Change-Id: I4a33464e8c0ba65570cfe670aad6bb755e6335c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/524384
Commit-Queue: Bryce Lee <brycelee@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: William Xiao <wxyz@google.com>