[ffx][keyboard] Adds the command `ffx keyboard`
This is an experimental command that allows us to configure the keymap
to use in the keyboard's input pipeline.  It only exists until the user
interface for the keyboard settings is implemented and  ready to go.

Removed the original `text_input_cli` since it is not possible to route
a request from a command line utility into a session without special
plumbing.  Instead, reusing all of that code in the context of an
experimental `ffx` command, as the session is accessible through the
`ffx` tool.

Bug: 76057, 76058, 76056
Change-Id: Ifc4bdea5bd7659a566bb8758157492913b21fc46
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530509
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Matthew Boetger <boetger@google.com>