[assembly] Allow clients to specify the bootloader name
The bootloader name is not always firmaware_<type>. In the current
build, in fact, it is almost never this format. To make comparison
between the current build and the image assembly outputs easier, we are
going to allow the name to be specified in the config.

Change-Id: Ia366d405ec0be9623f25e8e28b61b86de9a69f6a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/531754
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>