[blobfs] Reland host support for reading multi-extent blobs.
Reverted in commit e99b65248d73d7e7dec5e89883e19abecfa0b299

The missing piece was lack of support for zero length merkle trees.

This more closely unifies how the host and target read blocks for blobfs
but required some changing of the cache behaviour. The nodemap is now
held in memory as is done on the host so that it can use the same
libraries. This also meant changing the use of a dirty cache block for
flushing the inode. Instead the relevent block is just flushed directly
from the in-memory version, and since this was the only usage of the
dirty cache flag, it has been removed.

Tested: Manually.

Bug: 48571, 32953, 76514
Change-Id: Ica5eb9158798e0d9e76cee93c06e99e759062064
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/531925
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>