[netemul] Copy v1 client crate and common integration library
Also add allowlists for both libraries to prevent additional uses
accruing.

This will allow migration of the existing v1 client crate and library
of common integration test library to netemul-v2 in-place, so that git
history can be preserved; in the meantime, existing dependencies will be
on this copy, which can eventually be deleted.

Bug: 76475
Test: fx test -d netstack-integration-tests
Change-Id: I7659a8e023245769ea79533b76adf6514a457c00
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/531381
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>