[ffx][iquery] Make iquery code a rust library
This introduces a rust_library to the iquery codebase which the binary
then depends on. As part of this change I tried to get as much as
possible available on host as well as target. There are still some open
questions mostly around files where things are only running on target. I
introduced a test which then makes it possible to ensure that the host
side stuff compiles correctly.

There should be zero functional difference from this change so the
primary testing story here is that the existing tests continue to pass.

Change-Id: Icf5c768e5ccd55e1aaa1caae026c8a22d5134df4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525106
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>