[system-updater] Migrate system-updater to a v2 component.
This change migrates the system-updater component from the software
delivery stack into the core realm as part of the v1-v2 migration
project. System-updater is an eagerly started component that provides
access to the fuchsia.update.installer.Installer service.

As part of this migration the component depends on several services
provided by other components in the SWD stack. These are routed through
core.cml to the component. In addition to this this component relies on
a set of features & directories:
- build-info
- deprecated-misc-storage
- isolated-persistence-storage
- pkgfs/system

Bug: 73992
Change-Id: If9234f609c7259f24218ee2cc80a1020af7867ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515223
Reviewed-by: Dan Johnson <computerdruid@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>