[assembly] Generate salt file during GN-build
In order to compare the image assembler output with the GN-build output,
we need to use the same VBMeta salt. We are spitting out the GN-build's
salt to a file, so that it can be consumed by the image assembler to
produce consistent outputs.

Change-Id: I088be618d90ec34fba47eacdea9d92e4922819e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532381
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>