[kernel][vm] Evictor frees discardable pages
The discardable VMO side code collects the list of freed pages, and
defers the actual freeing of pages to the Evictor. This keeps the
discard APIs consistent with the eviction one, where pager backed VMOs
also rely on the Evictor to free evicted pages.

This change paves the way for unit tests in the following commit,
allowing us to verify exact free counts and freed pages. The tests are
added in a separate commit for ease of reviewing.

No change in existing functionality.

Bug: 72412
Change-Id: I2f7a1cf6cb7454f2bf11a589271e6f9ac7c4b40b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/524543
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>