[assembly] Disable passing boot args to ZBI
The boot args can be supplied via the IMAGE_ARGS field in the ZBI, or in
the devmgr_config.txt file. In order to enable binary diffing of the
GN-build ZBI, we are going to use the devmgr_config.txt for now.

Change-Id: Ibb6eb297f4a1df1c58577419b1cd86b99e3584fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532104
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>