[assembly] Remove whitespace from update_package.json
Removes the whitespace from the update packages' meta/package so that it
can be compared to the update package generated by the image assembler.
We are stripping the extension, so that `fx format-code` does not reject
the change.

Change-Id: Ib88e33f38ad5df3f09922e6ba0fc56ebea5445c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532075
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>