[goldfish-display] Remove references to layer images if layer is unused.
If we unset a layer in SetDisplayLayers() for a goldfish
display device, we should remove all references of images
used in that layer in the driver so that the layer change
in display controller could be safely applied on next Vsync
signal.

This fixes an issue on session restart where Scenic creates
and sets a new layer for the display.

Bug: 74677
Change-Id: Iffeb8c803893e5bcd49314a4c714020c1bdee439
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532025
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: David Worsham <dworsham@google.com>