[scenic][display] Remove DCHECK on unhandled vsync images.
For a certain DisplaySwapchain instance, it is possible that image
IDs in OnVsync() events doesn't match any of the images it imported
in the past; for example, it could be from other DisplayCompositor
using the same display controller FIDL connection. It is okay if we
simply ignore these images with "invalid" handles.

TEST: "fx shell session_control restart" on workstation.64 (no crash)
Bug: 74677
Change-Id: If5bc6b37bc25d84107f1f7b26a6a07fdd04b1133
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532245
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Worsham <dworsham@google.com>
Reviewed-by: Josh Gargus <jjosh@google.com>