[boot-options]: Reduce stack size for tests.
Couple of times we've blown up the stack, due to the size of the boot
options and the multiple instances in each test. Alternatives such us
using the bootloader require some allocator for deletion which is not
available for physboot. On that note, we use static storage for a single
instance of boot options, and one accessor method, that resets its
content.

Test: k ut boot-options
Change-Id: I2a8910c40283ac78906a16e30b5f347b8e225d42
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526161
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>