[carnelian] Fixed BufferLayerCache cloning.
The cleanup logic for BufferLayerCache was incorrect since it was
upgrading the weak pointer to the SmallBitSet before the last
clone was dropped. This is fixed by using the other Rc in the
object to know when only the last clone is dropped.

This CL also contains a subtle fix in the clear logic: the
SmallBitSet should not be touched when clearing; it was introduced
by mistake in the last fix.

Last but not least, SmallBitSet is now not Copy anymore to reduce
the risk of accidental clones.

While testing, I've also removed a couple of debug prints added
previously by mistake.

Testing: fx test mold-tests & button-rs & layout-gallery-rs
Change-Id: I477ccebb2f242b75882421bb07a19c084ae96be3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532321
Commit-Queue: David Reveman <reveman@google.com>
Reviewed-by: David Reveman <reveman@google.com>