[fidlc] Convert remaining failure tests
This updates the script to catch failure tests that don't exist
as a Foo/FooOld pair, and converts any remaining failure tests.
It also adds support for detecting difference in errors between
failure tests (e.g. if the fidlc returns error A in the old
syntax and error B in the new syntax), and fixes remaining tests
so that either 1. their errors match or 2. there is inline
documentation for why there's a behavior difference.

The only behavior change is to reorder validation code for handles
to provide a different (more natural) error.

Bug: 72924
Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler

Change-Id: If90256c2d15e2851952a5a0f9c86282c0e529cc7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532002
Commit-Queue: Felix Zhu <fcz@google.com>
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>