[vfs] Clarify recycling.
fbl::Recyclable is unusual because it bypasses most virtual inheritance
even though it looks virtual. This clarifies this behavior in the Vnode
comment.

Redirects the fbl_recycle() calls to a virtual function implemented on
fs::Vnode. This bypasses the unusual behavior of fbl_recycle() to a
single virtual function that is more predictable for users.

Critically PagedVnode lacked a recyclable derivation and function so
RefPtr<PagedVnode> would bypass the Blobfs cache.

Change-Id: Idecb12ad77497109ec6aa0e07c0f4fcc458b0a72
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532601
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>