[blobfs] Change ownership of paged vnodes.
Change the ownership model for paged vnodes. This moves the owning
reference that the PagedVfs used to keep to the PagedVnode itself. This
way, the PagedVnode can always be registered for page requests for as
long as the VMO is alive without affecting its lifetime. This avoids
some race conditions identified in the added comments.

Use a separate lock for the Vnode tracking to prevent reentrant locking.
This was happening because Open in the Vfs creates the node while the
Vfs is locked, and then the created node is calling back into the Vfs to
register itself.

Fix the StreamingWritesTest by using the new BlobfsTestSetup.

Add a commented-out test for detaching vmos in the new pager.

All blobfs unit tests pass with the new pager with this patch.

Bug: 76538
Change-Id: If38fa46bdf6ea77bcaa5a57e58eae45edfdd25a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530963
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>