[magma] Add ICD manifest json to package
In the future the loader service will read the manifest json from the
package and pass it to libvulkan.so in the client. This changes adds a new resource
target for that manifest to the icd_config template and updates the
Intel and Goldfish ICDs to use it.

Change-Id: I14f7708ed0972a31b74c07a43b9b7be35418be0f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/531681
Commit-Queue: John Bauman <jbauman@google.com>
Reviewed-by: Craig Stout <cstout@google.com>
Reviewed-by: John Rosasco <rosasco@google.com>