[vulkan] Add feature query method to loader
The vulkan loader service in an image could be built at an earlier or
later time than libvulkan.so in a package. To help newer versions of
libvulkan.so deal with older versions of the loader service, add a
GetSupportedVersion method to query loader service features.

To help newer versions of the loader service deal with older,
potentially buggy versions of the client, add a ClientVersion enum to
the method. Currently the value will always be set to 0, but if bugs in
the client are found and the loader needs to work around them we can
increment the version number.

Change-Id: I5f1828b597e56db77e3cd21b928d967d96112d6a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532162
Reviewed-by: Dale Sather <dalesat@google.com>
Reviewed-by: Craig Stout <cstout@google.com>
API-Review: Dale Sather <dalesat@google.com>
Commit-Queue: John Bauman <jbauman@google.com>