[boot-options]: Kernel test is not thread hostile.
If kernel-test is compiled in kernel environment, it will use
ktl::unique_ptr, while if its compile in for phys environment it will be
static storage on the bss section.

Now this is hidden by a getter that provides pointer-like interface, so
tests only need to assume they have some sort of pointer. For static
storage, there is nothing to do, for ktl::unique_ptr it gets destroyed
when it goes out of scope.

To keep it simple, allocation failure, forces the getter to return
nullptr, and the caller asserts they got a non null ptr.

Bug: 77209 # Done
Test: k ut boot-options
Change-Id: I16e1adc4f9396a587d229a24e169b9654b0470c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533101
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>