[cm][routing] Provide capability security policies to RoutingTestModel
Adds a method to the RoutingTestModelBuilder trait allowing
the caller to set the primary and debug capability security
policies in the test model's RuntimeConfig.

Moves the routing tests of routing that involve these policies
into CommonRoutingTest and runs them for both ComponentManager
and the static routing analyzer.

Bug: 61861
Change-Id: Id2e02c65ec496c7080edb6c8312ec4dc19d1a5d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/524549
Commit-Queue: Laura Peskin <pesk@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>