[fx] Process all command line args for `fx emu`
Previously, when calling fx emu, if the script hit a command-line
argument it didn't recognize, it assumed all following args were meant
for the emulator executable, so it stopped processing args and passed
all remaining args to the emulator. This wasn't documented anywhere, and
isn't a very intuitive approach.

This CL adjusts the processing code so that any args the startup script
doesn't recognize will be set aside to be passed to the emulator, but
the processing will continue until all args have been processed.

Bug: https://fxbug.dev/47574
Change-Id: I566b0613d7c5d3d6ed49734f5b2452bab35d889a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/527780
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>