[boot-options]: Fix unknown value handling.
Before this patch, running into an unknown value, would be a no-op
operation. This is an undesired behavior change from previous command
line library. The previous command line library would return the last
entry for a given option, no matter whether it was valid or not, since
the validation was delagated to the library user.

This patch introduces a default value for each member variable, and
updates Parse to return true when it successfully parses a value and
false when it doesnt.

Bug: 53594
Test: boot-option-tests
Change-Id: I7c382ff3433220173bb1e69baa5850a7377313ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526084
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>