[audio][lib] Update WavWriter for packed-24
The WavWriter library has been updated to accept SIGNED_24_IN_32
streams, but write packed-24 streams to file if the bits_per_sample
parameter to Initialize is set to 24. If it is set to 32, then the
library will write the 24-in-32 stream directly to disk as provided
to it by the client.

Bespoke code to support packed-24 has been removed from the utility
wav_recorder, as clients no longer need implement this themselves.

Parameterized unit tests are added, to validate file sizes and
sample sizes (both in-stream and in-file).

fx test -o audio-libwav-unittests

Change-Id: I85582b005585476f586648f1ab74aed200f21758
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530200
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>