[scripts][clang] Write script output to file
THe clang canary builders were failing because this script could run
multiple times without producing a no-op in the build. This updates the
script to accept a path fo a file and write to that file rather than
dumping to stdout.

This also updates the action that onvokes this script to write to an
file in the build directory.

Change-Id: Ib39d8a9cd489837f7793be5dda0c925695d5dcc0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532981
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>