[debugger] Fall back to binary names when local.
When debugging on the local system, look for symbols in the binary
itself (based on the local path) if the build ID isn't found in the
normal index.

Currently this is not used but will enable local debugging in the
future. The detection is currently primitive (looks for "localhost") but
as we enable this mode local debugging will likely not require a
separate debug agent so it can be a direct in-process connection.

Change-Id: I856d92de7c807f325df8637662d972296ba1e25c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/530146
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>