[fuchsia-async] Make duration work on host
This change exposes Duration as a type from fuchsia-async the same way
that Time is exposed. This allows for platform agnostic timeouts to be
used.

The reason for this change is to let readdir_recursive work on host.
That is also included in this CL to help motivate this change.

Change-Id: I77051c63cae7ed6781f01924944796abf2695cab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532901
Reviewed-by: James Tucker <raggi@google.com>
Commit-Queue: Andy Weiss <dragonbear@google.com>