[fx] Add subcommand "remote" to fx vdl and fvdl
As part of the new FEMU DX initiative, we're working towards unifying
how people use the emulator. This CL starts the process of adding a new
`remote` subcommand to the `fx vdl` and `fvdl` tools, which mirrors the
existing `fx emu-remote` functionality.

This CL introduces a placeholder subcommand, as well as the args that
will need to be supported by that subcommand. The subcommand is not
functional at this time, so we put a message in the help text indicating
that it's only a placeholder.

Change-Id: I7fc4bf2ad30745fb716fbbcef6d2b247c76ef7cc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532582
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Yuan Zhi <yuanzhi@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>