[kernel][lockdep] Fix a HW WDT triggered by lockdep during shutdown.
While the details which lead to the lockup are complicated (see
fxb/77097), the short version is that the existing mechanism for
triggering the loop detector thread is not safe in all situations.
Specifically, the detector was attempting to signal a kernel event,
but could be doing so when preemption had not been explicitly
disabled, and while inside of a non-thread-lock spinlock.

If this happens, and:

1) The event was not currently signaled.
2) The loop detector thread was currently blocked on the signal.
3) The current CPU is chosen for the loop detector thread as a result
   of the signal operation.

Then we will end up immediately triggering a preempt operation on our
current core, which will land us in RescheduleCommon while holding a
spinlock which cannot be held across a context switch (which then
triggers an ASSERT).

For now, in order to avoid this situation, I'm simply removing the
event entirely and switching to a model where the trigger simply sets
a "dirty" flag, and the loop checker thread polls this flag using its
existing throttle rate of once every 2 seconds.

It is possible to go back to the event based system, but we would need
to address the issue of how to properly defer preemption while holding
a non-thread-lock spinlock.  There are a few options here, but each of
them is something of a significant change from our current model, and
should be discussed thoroughly before any changes are made.

Since the violation which leads to the ASSERT and eventually the
lockup during shutdown can _technically_ happen during normal
operation (although, it is terribly unlikely), I think it is important
to put a mitigation in place immediately, before moving on to talk
about preempt-disable behavior re: spinlocks.  If we get to a point
where we can properly handle deferring preemption in the situation
described above, we can come back and revert this CL.

Fixed: 77097
Change-Id: I95f4c42ff9c7ee43b3cfac8f1446c8c59439f217
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533642
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>