[debugger] Track calling convention on structures.
DWARF defines a calling convention for structures. This is used to
figure out how to pass and return values.

Read this from the symbol file ane track on the Collection symbol
object.

This information is displayed in sym-info but is not used for any
functional reason. It will be needed to display return values from
functions.

Bug: 77333
Change-Id: I5f52d73eeb2d132da1a039dfbc0dfac1baa225fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532764
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>