[session] Add support for Atlas 4K display resolution
This change adds support for the Atlas 4k resolution.

It re-purposes an existing 4k 27" monitor resolution because Rust does
not allow defining additional monitor metrics with same resolution as
existing definitions. Since we need to support Atlas 4K officially we
are dropping the values for the 27" 4K arbitrary monitor.

This mechanism is anyways too crude to support arbitrary monitors. We
will need better solution to allow for that.

The magic value to scale the resolution on Atlas 4K was chosen such that
flutter reports physical display size of 3840x2160 and display pixel
ratio of 2.0

Change-Id: If1a2cd589eb4071ea16f91ac5a23875d64c3f309
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533982
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Sanjay Chouksey <sanjayc@google.com>