[ffx] Convert target formatter tests to goldens
As part of https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=74157 I
am going to change some of the format of the output of target list.
These tests have a bunch of hardcoded formatted strings and a TODO to
replace those with goldens. This does that transition to use golden
files. The functionality should be the same so the goal here is that the
current tests continue to pass with this change.

Change-Id: I4a83b02d987a02c43363a5ea1280bbf23fc71997
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533302
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>