[covargs] Merge entries from all symbolizer output files.
Multiple files could contain entries with the same name (which is the
name of one of the profraw files) but with a different set of modules
and segments. These would get overwritten with whatever's in the last
file that references that profraw file which would cause coverage
collected from previous files to be missing if the same modules weren't
seen anywhere else.

Bug: 77034
Change-Id: Ideb5c30c9a48c5464d900b04429b2453e4262ad6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534161
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>