[soundplayer] Route LogSink
Follow up to fxr/504523.
Noticed this because I saw an error log about soundplayer using logsink
but not being offered it.

It's not apparent in fxr/504523 that soundplayer uses LogSink and
therefore is missing a route because the use of LogSink is slightly
obscured by the fact that it's done in the logsink manifest shard.
This specific issue is not covered in the migration guide.
I'm not sure if ffx scrutiny would have caught this or not. This may be
shadowed by fxbug.dev/77336 (speculating).

How can we do better in the future?

Change-Id: Ic609feac2006c7df18118d3ec7aa3abd8315f017
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533044
Fuchsia-Auto-Submit: Shai Barack <shayba@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>
Commit-Queue: Shai Barack <shayba@google.com>