[ffx] Change target list to have multiline IPs
This introduces the concept of a field that can be either a string or an
vector of strings. If it is a vector of strings then each string is
printed on it's own line. This shrinks the size of that array column to
the maximum string length over the array.

https: //bugs.fuchsia.dev/p/fuchsia/issues/detail?id=74157
Change-Id: I5c254b93db385306cfa78fb6d2593dabe9d3aba4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533303
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>