[gidl][rust] Add handle rights coverage in rust
This CL covers handle rights logic in rust. It does so by adding new
GIDL tests, NOT by adding deep equals checking for cases where handles
might change value. The tests that do that are still disabled, but
largely redundant for the purpose of testing rust. Work on that is
deferred until a future point where it is necessary.

Bug: 41920
Change-Id: I3b92c3a8af30bce86a47d4b65badb7799a5a9bb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533822
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>