[vfs] Use a separate lock for tracking vnodes.
As with the PagedVfs tracking of registered vnodes, the Vfs tracks open
nodes (this duplicate tracking can hopefully go away when the paging
capability is completely landed and merged with everything).

And just was with the PagedVfs, the Vfs needs a separate lock for
tracking opened nodes to avoid reentrant locking (nodes are destroyed
and created from within the main Vfs lock, so they can't call back into
the Vfs to register with the same lock).

This makes most of the integration tests pass and allows the system to
write new blobs at runtime (previously this would fail so no tests
could be run on a system with the new pager enabled).

Bug: 51111
Change-Id: Iea15214159aa622f7131a593291759c263527951
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533262
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>