[omaha-client] Fix flakes when subtracting instants
It's possible that `Instant::now()` is the number of monotonic seconds
after a device is booted. This means that subtracting a value from
`Instant::now()` can risk an underflow panic, especially if the test is
run on a device that was just booted.

This patch fixes a flake in omaha-client by making sure that whenever we
subtract a value, we first advance the `Instant` value by 24 hours
before subtracting from it to make sure we don't trip over this
underflow.

Change-Id: Idb6bf55cebe2755e464cee00f84114db8f996ff3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534522
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>