[rive-rs] Removed allocating asserts.
Due to an oversight, I have introduced a few Option expect calls
that do heap allocations regardless of the state of the assert. The
anti-patterns looks like this:

    option.expect(&format!("{}", ...))

Looking at this now, it seems unfortunate that the 'static lifetime
bound wasn't enforced on the definition of Option::expect.

Change-Id: I94b10c2a5f7dbe436a4cf6cd54547b880c51506d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534583
Reviewed-by: David Reveman <reveman@google.com>
Commit-Queue: David Reveman <reveman@google.com>